0.1.1
Messed up with some methods, File browsing for packagestore.manifest only allows file selection with that exact name, it's come to my attention that the file may not be that exact name.

Same goes for IoStoreCommands.txt

Added a method for loading UnrealPakConfig when IoStore Mode is toggled and the repack button is pressed.